Modified OBD2Kit for use with Mac OS X; Removed support for GoLink device; Added support for serial devices; Logging of NSIntegers types now cast to long; Separated Debug/Info logging defines; Added init timer to ELM327 classes; Removed Core Location code;

+++ Classes/device_drivers/elm327/ELM327Command.m
@@ -48,10 +48,10 @@ NSString *const kELM327SetDeviceIdentifier			= @"AT @3";
 	ELM327Command* cmd = nil;
 	
 	if (pid >= 0x00 && pid <= 0x4E) {
-		cmd = [[ELM327Command alloc] initWithCommandString:[NSString stringWithFormat:@"%02x %02x", (NSUInteger)mode, pid]];	
+		cmd = [[ELM327Command alloc] initWithCommandString:[NSString stringWithFormat:@"%02x %02x", mode, (unsigned int)pid]];
 	}
 	else {
-		cmd = [[ELM327Command alloc] initWithCommandString:[NSString stringWithFormat:@"%02x", (NSUInteger)mode]];	
+		cmd = [[ELM327Command alloc] initWithCommandString:[NSString stringWithFormat:@"%02x", mode]];
 	}
 
 	if(data) {
